Subscriptions_admin_functionality

+++ react-resolution-super/src/components/subscriptionCard/index.tsx
@@ -1,37 +1,97 @@
-import React, { useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import * as ST from './styled'
 import { IPatchSubscription, ISubscriptionWithId } from 'types/subscription'
 import { deleteSubscription } from 'api/subscription'
 import DefaultPopup from '../ui/Modals/defaultModal'
-import { ConfirmDelete, ModalEditSubscription } from './ModalEitSubscription'
+import {
+  ConfirmDelete,
+  ModalEditSubscription,
+  ConfirmUnsubscribe,
+  ResultUnsubscribe,
+} from './ModalEitSubscription'
+import { Payment } from '../views/payment'
+// @ts-ignore
+import { HunelProvider, HunelCreditCard } from 'reactjs-credit-card'
+import InformModal from '../ui/Modals/InfromModal'
+import { setActivePurchase } from '../../store/actions/purchase'
+import { useDispatch } from 'react-redux'
 
 interface ICard {
-  key: number
+  key?: number
   props: ISubscriptionWithId
   isAdmin: boolean
-  deleteSubscriptionFromList: (idSubscription: number) => void
-  updateSubscriptionList: (updatedSubscr: IPatchSubscription) => void
+  deleteSubscriptionFromList?: (idSubscription: number) => void
+  updateSubscriptionList?: (updatedSubscr: IPatchSubscription) => void
+  unsubscribe?: () => void
+  responseModalText?: string
+  showResponse?: boolean
+  setResponseModalText?: React.Dispatch<React.SetStateAction<string>>
+  isPaid?: boolean
+  onProlong?: () => Promise<any>
+  update?: React.Dispatch<React.SetStateAction<string>>
 }
 
-export const SubscriptionCard = ({
+const SubscriptionCard = ({
   props,
   isAdmin,
   deleteSubscriptionFromList,
   updateSubscriptionList,
+  unsubscribe,
+  responseModalText,
+  showResponse,
+  setResponseModalText,
+  isPaid,
+  onProlong,
 }: ICard) => {
+  const hunel = new HunelCreditCard()
+  const dispatch = useDispatch()
   const [showDeleteModal, setShowDeleteModal] = useState<boolean>(false)
   const [showEditModal, setShowEditModal] = useState<boolean>(false)
+  const [showUnsubscribeModal, setUnsubscribeModal] = useState<boolean>(false)
+  const [showPayModal, setPayModal] = useState<boolean>(false)
+  const [showResultModal, setResultModal] = useState<boolean | undefined>(false)
+  const [showSubmitModal, setShowSubmitModal] = useState<boolean>(false)
+
+  const dispatchProlong = () => {
+    dispatch(setActivePurchase())
+  }
+
+  useEffect(() => {
+    setResultModal(showResponse)
+  }, [showResponse])
+
   const handleDeleteModal = (): void => {
     setShowDeleteModal(!showDeleteModal)
   }
 
+  const handleResultModal = (): void => {
+    setResultModal(false)
+    if (setResponseModalText) {
+      setResponseModalText('')
+    }
+  }
+
   const deleteContactItem = (idSubscription: number): void => {
-    deleteSubscription(idSubscription).then((resp) =>
-      deleteSubscriptionFromList(idSubscription)
-    )
+    if (deleteSubscriptionFromList) {
+      deleteSubscription(idSubscription).then((resp) =>
+        deleteSubscriptionFromList(idSubscription)
+      )
+    }
   }
   const handleEditModal = (): void => {
     setShowEditModal(!showEditModal)
+  }
+
+  const handleUnsubscribeModal = (): void => {
+    setUnsubscribeModal(!showUnsubscribeModal)
+  }
+
+  const handlePayModal = (): void => {
+    setPayModal(!showPayModal)
+  }
+
+  const handleModal = (): void => {
+    setShowSubmitModal(!showSubmitModal)
   }
 
   return (
@@ -60,6 +120,18 @@ export const SubscriptionCard = ({
             <ST.InfoItem>{'All coefficients are available'}</ST.InfoItem>
             <ST.InfoItem>{props.description}</ST.InfoItem>
           </ST.SubscriptionInfoList>
+          <ST.ActionsButtonsContainer>
+            {!isPaid && (
+              <ST.ProlongButton onClick={() => handlePayModal()}>
+                Pay to prolong
+              </ST.ProlongButton>
+            )}
+            {!isAdmin && (
+              <ST.UnsubscribeButton onClick={() => handleUnsubscribeModal()}>
+                Unsubscribe
+              </ST.UnsubscribeButton>
+            )}
+          </ST.ActionsButtonsContainer>
         </ST.CardBody>
       </ST.SubscriptionCard>
       {isAdmin && (
@@ -89,6 +161,56 @@ export const SubscriptionCard = ({
           onClose={handleDeleteModal}
         />
       )}
+
+      {unsubscribe && (
+        <DefaultPopup
+          children={
+            <ConfirmUnsubscribe
+              onUnsubscribe={() => unsubscribe()}
+              closeModal={() => handleUnsubscribeModal()}
+            />
+          }
+          show={showUnsubscribeModal}
+          onClose={handleUnsubscribeModal}
+        />
+      )}
+      {showResultModal && (
+        <DefaultPopup
+          children={
+            <ResultUnsubscribe
+              responseModalText={responseModalText}
+              closeModal={() => handleResultModal()}
+            />
+          }
+          show={showResultModal}
+          onClose={handleResultModal}
+        />
+      )}
+      {showPayModal && (
+        <DefaultPopup
+          children={
+            <HunelProvider config={hunel}>
+              <Payment
+                closeModal={() => handlePayModal()}
+                subscriptionInfo={props}
+                onSubmitPay={onProlong}
+                setShowSubmitModal={setShowSubmitModal}
+                showSubmitModal={showSubmitModal}
+                dispatchFunction={dispatchProlong}
+              />
+            </HunelProvider>
+          }
+          show={showPayModal}
+          onClose={handlePayModal}
+        />
+      )}
+      <InformModal
+        text={'Thanks for your payment!'}
+        show={showSubmitModal}
+        onClose={handleModal}
+      />
     </>
   )
 }
+
+export default SubscriptionCard
